+++ packages/bpk-component-image/index.js
@@ -17,12 +17,14 @@
  */
 
 import BpkImage from './src/BpkImage';
+import BpkBackgroundImage from './src/BpkBackgroundImage';
 import withLazyLoading from './src/withLazyLoading';
 import withLoadingBehavior from './src/withLoadingBehavior';
 
 export default BpkImage;
 
 export {
+  BpkBackgroundImage,
   withLazyLoading,
   withLoadingBehavior,
 };