+++ changelog.md
@@ -7,8 +7,8 @@ _Nothing yet..._
 ## 2017-08-30 - Calendars/Datepickers can now set initally focused date / month
 
 **Added:**
- - bpk-component-calendar: 4.0.17 => 4.1.0
- - bpk-component-datepicker: 7.0.18 => 7.1.0
+- bpk-component-calendar: 4.0.17 => 4.1.0
+- bpk-component-datepicker: 7.0.18 => 7.1.0
   - Added `initiallyFocusedDate` prop that allows controlling which date is initially focused when `selectedDate` is not specified. If not supplied `minDate` will receive focus as before.
 
 ## 2017-08-30 - New image component